[#3558] StringLiteralCorrector extraction
