+++ lib/rubocop.rb
@@ -145,6 +145,7 @@ require_relative 'rubocop/cop/correctors/multiline_literal_brace_corrector'
 require_relative 'rubocop/cop/correctors/ordered_gem_corrector'
 require_relative 'rubocop/cop/correctors/parentheses_corrector'
 require_relative 'rubocop/cop/correctors/punctuation_corrector'
+require_relative 'rubocop/cop/correctors/string_literal_corrector'
 
 require_relative 'rubocop/cop/bundler/duplicated_gem'
 require_relative 'rubocop/cop/bundler/insecure_protocol_source'
