+++ lib/rubocop/cop/mixin/string_literals_help.rb
@@ -6,6 +6,8 @@ module RuboCop
     module StringLiteralsHelp
       include StringHelp
 
+      private
+
       def wrong_quotes?(node)
         src = node.source
         return false if src.start_with?('%', '?')
@@ -13,19 +15,6 @@ module RuboCop
           !double_quotes_required?(src)
         else
           src !~ /" | \\[^'] | \#(@|\{)/x
-        end
-      end
-
-      def autocorrect(node)
-        return if node.dstr_type?
-
-        lambda do |corrector|
-          str = node.str_content
-          if style == :single_quotes
-            corrector.replace(node.source_range, to_string_literal(str))
-          else
-            corrector.replace(node.source_range, str.inspect)
-          end
         end
       end
     end
