+++ lib/rubocop/cop/style/string_literals_in_interpolation.rb
@@ -23,6 +23,10 @@ module RuboCop
         include ConfigurableEnforcedStyle
         include StringLiteralsHelp
 
+        def autocorrect(node)
+          StringLiteralCorrector.correct(node, style)
+        end
+
         private
 
         def message(*)
